+++ examples/aggregation_example.py
@@ -1,4 +1,4 @@
-from pious.pio.aggregation import AggregationReport
+from pious.pio.pio_aggregation_report import AggregationReport
 from pious.pio.resources import get_aggregation_root, get_database_root
 from ansi.color import fg
 from ansi.colour.fx import italic
